backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
build==1.5.0
certifi==2026.4.22
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
fastapi==0.136.1
filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.4.0
googleapis-common-protos==1.74.0
greenlet==3.5.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.13.0
idna==3.13
importlib-metadata==8.7.1
importlib-resources==7.1.0
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.17
langchain-chroma==1.1.0
langchain-classic==1.0.5
langchain-community==0.4.1
langchain-core==1.3.2
langchain-mistralai==1.1.3
langchain-openai==1.2.1
langchain-pinecone==0.2.13
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.8.0
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy-extensions==1.1.0
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.25.1
openai==2.34.0
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pillow==12.2.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.4.1
protobuf==6.33.6
pybase64==1.4.3
pydantic==2.13.3
pydantic-core==2.46.3
pydantic-settings==2.14.0
pygments==2.20.0
pypdf==6.10.2
pypika==0.51.1
pyproject-hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
pyyaml==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.23.1
tqdm==4.67.3
typer==0.25.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.46.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
